DiscountCodeInfo as interface #1053
